Adding proc tools for the second pass inc langaus fits, and factorising the NN module a bit more
