+++ BetaDiscTool/full_analysis_sndisc/BATRA-SNDisc/secondPass_ProcTools.py
@@ -1,0 +1,35 @@
+import numpy as np
+import matplotlib.pyplot as plt
+import pandas as pd
+
+from landaupy import langauss
+from scipy.stats import median_abs_deviation
+from scipy.optimize import curve_fit
+
+def binned_fit_langauss(samples, bins, min_x_val, max_x_val, channel, nan='remove'):
+  if nan == 'remove':
+    samples = samples[~np.isnan(samples)]
+
+  hist, bin_edges = np.histogram(samples, bins, range=(min_x_val,max_x_val), density=True)
+  bin_centres = bin_edges[:-1] + np.diff(bin_edges) / 2
+  hist = np.insert(hist, 0, sum(samples < bin_edges[0]))
+  bin_centres = np.insert(bin_centres, 0, bin_centres[0] - np.diff(bin_edges)[0])
+  hist = np.append(hist, sum(samples > bin_edges[-1]))
+  bin_centres = np.append(bin_centres, bin_centres[-1] + np.diff(bin_edges)[0])
+
+  hist = hist[1:-1]
+  bin_centres = bin_centres[1:-1]
+  landau_x_mpv_guess = bin_centres[np.argmax(hist)]
+  landau_xi_guess = median_abs_deviation(samples) / 5
+  gauss_sigma_guess = landau_xi_guess / 10
+
+  popt, pcov = curve_fit(
+    lambda x, mpv, xi, sigma: langauss.pdf(x, mpv, xi, sigma),
+    xdata=bin_centres,
+    ydata=hist,
+    p0=[landau_x_mpv_guess, landau_xi_guess, gauss_sigma_guess],
+  )
+  return popt, pcov, hist, bin_centres
+
+def gaussian(x, A, mu, sigma):
+  return A * np.exp(-0.5 * ((x - mu) / sigma) ** 2)
